+++ sites/demo/CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
+- Include version in CMS cache prefix to bust cache when deploying new version
 - Set default language to english (also makes generating the demo site work)
 
 ## [1.3.0] - 2020-12-07